# openspec/versions/v0.0.1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/groups/{id}/members/{memberId}": {
      "delete": {
        "description": "Owner remueve un miembro del grupo",
        "operationId": "GroupsController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Miembro removido"
          },
          "403": {
            "description": "No eres el owner"
          },
          "404": {
            "description": "Miembro no encontrado"
          }
        },
        "summary": "HU: Sacar miembro del grupo (Solo owner)",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/messages/membership/{id_membership}": {
      "get": {
        "operationId": "MessagesController_findByMembership",
        "parameters": [
          {
            "name": "id_membership",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensajes de la membresía."
          }
        },
        "summary": "Obtener mensajes de una membresía",
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_getUnreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/expo-push-token": {
      "post": {
        "operationId": "NotificationsController_saveExpoPushToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpoPushTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/expo-push-token/{token}": {
      "delete": {
        "operationId": "NotificationsController_deleteExpoPushToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/preferencias": {
      "get": {
        "operationId": "NotificationsController_obtenerPreferencias",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_actualizarPreferencia",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenciaCanalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/memberships": {
      "post": {
        "operationId": "MembershipsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      },
      "get": {
        "operationId": "MembershipsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      }
    },
    "/api/memberships/{id}": {
      "get": {
        "operationId": "MembershipsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      },
      "patch": {
        "operationId": "MembershipsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      },
      "delete": {
        "operationId": "MembershipsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      }
    },
    "/api/memberships/user/{id_user}": {
      "get": {
        "operationId": "MembershipsController_findByUser",
        "parameters": [
          {
            "name": "id_user",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      }
    },
    "/api/memberships/group/{id_group}": {
      "get": {
        "operationId": "MembershipsController_findByGroup",
        "parameters": [
          {
            "name": "id_group",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      }
    },
    "/api/memberships/{id}/toggle-admin": {
      "patch": {
        "operationId": "MembershipsController_toggleAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Memberships"
        ]
      }
    },
...
  },
  "info": {
    "title": "Uniconnect API",
    "description": "Documentación oficial de la API REST de Uniconnect — red social universitaria.\n\nTodos los endpoints protegidos requieren un JWT Bearer token obtenido en `POST /api/auth/callback`.\n\n**OpenAPI JSON**: `GET /docs-json`  |  **OpenAPI YAML**: `GET /docs-yaml`",
    "version": "0.0.1",
    "contact": {
      "name": "Uniconnect Team",
      "url": "",
      "email": "luma1017hl@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8007",
      "description": "Desarrollo local"
    },
    {
      "url": "https://uniconnect-backend.fly.dev",
      "description": "Producción (Fly.io)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ExpoPushTokenDto": {
        "type": "object",
        "properties": {}
      },
      "PreferenciaCanalDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateMembershipDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMembershipDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}